refactor: disconnect in p2psrv
